Complete refactor to standardize around 'LiveQueue's
A LiveQueue is a queue specially design to pull data from a sequeunce of
GeneralizedTrajectories. It itself is a GeneralizedTrajectories, so for the
vast majority of cases it should be treated as a single element. Should you
try to access a time past the end, the last valid value is returned
instead.

A GeneralizedTrajectory is simply a class that implements two function:
positionAtTime and speedAtTime. In some specific cases, the second can be
omitted, though the appropriate flag (speed_enabled = false) if it is used
in a LiveQueue.

By coallescing all management of sequences of data into a single class,
many difficult bugs were eliminated while significantly improving the logic
and hence the maintainability of the existing code.

Additionally, the code was rewritten to be modular around the main
components of the control structure and the entire codebase was restructed
into submodules.

Signed-off-by: Lukas Zhornyak <CainRose@users.noreply.github.com>